% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOExpenditures.R
\name{calcFAOExpenditures}
\alias{calcFAOExpenditures}
\title{calcFAOExpenditures}
\usage{
calcFAOExpenditures()
}
\value{
dataframe of FAO FBS
}
\description{
read gets FAO food expenditures
}
\author{
David M Chen
}
