% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMarkup.R
\name{calcMarkup}
\alias{calcMarkup}
\title{calcMarkup}
\usage{
calcMarkup(plot = FALSE)
}
\arguments{
\item{plot}{plot markups}
}
\value{
dataframe of ICP expenditures
}
\description{
calculate markups
}
\author{
David M Chen
}
