% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOPrices.R
\name{readFAOPrices}
\alias{readFAOPrices}
\title{readFAOPrices}
\usage{
readFAOPrices(fillPrices = T, level = "k")
}
\arguments{
\item{fillPrices}{true tries to fill prices with 5 year average around the missing value}

\item{level}{k or prim}
}
\value{
dataframe of FAO expenditures
}
\description{
read gets FAO food expenditures
}
\author{
David M Chen
}
