% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFBSnew.R
\name{readFBSnew}
\alias{readFBSnew}
\title{readFBSnew}
\usage{
readFBSnew(level = "k")
}
\arguments{
\item{level}{k prim or LEM}
}
\value{
dataframe of FAO FBS
}
\description{
read gets FAO food expenditures
}
\author{
David M Chen
}
