% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readICPexchangeRates.R
\name{readICPexchangeRates}
\alias{readICPexchangeRates}
\title{readICPexchangeRates}
\usage{
readICPexchangeRates()
}
\value{
dataframe of ICP expenditures
}
\description{
read ICP exchange rates
}
\author{
David M Chen
}
