% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressFAFH.R
\name{regressFAFH}
\alias{regressFAFH}
\title{regressFAFH}
\usage{
regressFAFH(
  weight = TRUE,
  plot = FALSE,
  predict = "SSP2",
  threshold = 0.85,
  calibrate = TRUE
)
}
\arguments{
\item{weight}{weighted by population regression}

\item{plot}{plot the line}

\item{predict}{SSP income scenario along which to predict}

\item{threshold}{upper threshold}

\item{calibrate}{calibrate lines for existing datapoints}
}
\value{
regression values for food away from home
}
\description{
calculate markups
}
\author{
David M Chen
}
