% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressMarkups.R
\name{regressMarkups}
\alias{regressMarkups}
\title{regressMarkups}
\usage{
regressMarkups(plot = TRUE)
}
\arguments{
\item{plot}{plot the regression lines}
}
\value{
dataframe of ICP expenditures
}
\description{
regress markups
}
\author{
David M Chen
}
