% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRedist.R
\name{toolRedist}
\alias{toolRedist}
\title{toolRedist}
\usage{
toolRedist(x, redist, output)
}
\value{
redistributed dataset
}
\description{
redistirbute other foods and catering categories
}
\author{
David M Chen
}
