% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valFAOExpenditures.R
\name{valFAOExpenditures}
\alias{valFAOExpenditures}
\title{valFAOExpenditures}
\usage{
valFAOExpenditures(plot = TRUE)
}
\value{
dataframe of FAO FBS
}
\description{
uses regression markups to calculate consumer price markups
}
\author{
David M Chen
}
