% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFBSnew.R
\name{readFBSnew}
\alias{readFBSnew}
\title{readFBSnew}
\usage{
readFBSnew(level = "k", return = "FBS")
}
\arguments{
\item{level}{k prim or LEM}
}
\value{
either return the regular FBS, or the proc_sugar and proc_oil datasheets
(important for weighting other functions)

dataframe of FAO FBS
}
\description{
read gets FAO food expenditures
}
\author{
David M Chen
}
